+++ src/common/index.ts
@@ -1,12 +1,11 @@
 export * from './apibase';
-export * from './constants';
 export * from './credentials';
 export * from './input';
 export * from './jrpcapi';
+export * from './keychain';
 export * from './nbytes';
 export * from './output';
-export * from './payload';
-export * from './persistenceoptions';
 export * from './restapi';
+export * from './secp256k1';
 export * from './tx';
 export * from './utxos';
