--- src/keychains/index.ts
@@ -1,2 +1,0 @@
-export * from './keychain';
-export * from './secp256k1';
