+++ src/utils/index.ts
@@ -1,4 +1,7 @@
 export * from './base58';
 export * from './bintools';
+export * from './constants';
 export * from './db';
 export * from './helperfunctions';
+export * from './payload';
+export * from './persistenceoptions';
